+++ packager/app/retired_flags.h
@@ -14,3 +14,6 @@ DECLARE_string(playready_key_id);
 DECLARE_string(playready_key);
 DECLARE_bool(mp4_use_decoding_timestamp_in_timeline);
 DECLARE_int32(num_subsegments_per_sidx);
+DECLARE_bool(generate_widevine_pssh);
+DECLARE_bool(generate_playready_pssh);
+DECLARE_bool(generate_common_pssh);